data gets now parsed again, emoved unnecessary uses
