+++ src/input/mod.rs
@@ -4,17 +4,14 @@ pub mod util;
 extern crate clap;
 
 use env;
-use log::LogLevelFilter;
 use std::convert;
 use std::fs::File;
 use std::io::{self, BufReader, BufRead};
 use std::iter::FromIterator;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
-use structs::flower::Flower;
-use structs::{self, Data};
 
-use self::clap::{App, AppSettings, Arg, ArgMatches, SubCommand};
+use self::clap::{App, AppSettings, Arg, SubCommand};
 
 
 /// Reads the arguments given to this program at execution and returns them
@@ -64,41 +61,6 @@ pub fn read_arguments<T>() -> config::GlobalConfig<T>
     config::GlobalConfig::from_arguments(matches)
 }
 
-// TODO: ueberbrueckungsmethode
-fn do_lame_stuff(matches: ArgMatches)
-                 -> (Result<Vec<Flower>, io::Error>, String, String, Option<LogLevelFilter>) {
-    let verbosity = match matches.value_of("verbosity") {
-        Some("debug") => Some(LogLevelFilter::Debug),
-        Some("info") => Some(LogLevelFilter::Info),
-        Some("error") => Some(LogLevelFilter::Error),
-        Some("off") => Some(LogLevelFilter::Off),
-        _ => None,
-    };
-    let data = parse_data(matches.value_of("data").unwrap());
-
-    (data,
-     "data/config.json".to_string(),
-     matches.subcommand_name().unwrap().to_string(),
-     verbosity)
-}
-
-// TODO: put this function somewhere else. Perhaps in main.rs?!
-fn do_awesome_stuff(matches: ArgMatches) {
-    // let data = ...
-    match matches.subcommand_name() {
-        Some("learn") => {
-            unimplemented!();
-            // let learn_params = ...
-            // learn(data, learn_params);
-        }
-        Some("classify") => {
-            unimplemented!();
-            // classify(data)
-        }
-        None => unreachable!(),
-        _ => unreachable!(),
-    }
-}
 
 /// reads content of given file and returns a result with
 /// either the Vector of Flowers or Err
@@ -113,14 +75,7 @@ pub fn read_data<T>(filename: &Path) -> io::Result<Vec<T>>
     reader.lines().map(|l| l?.parse::<T>()).collect()
 }
 
-// TODO: DOK
-pub fn into_data_vec(input: Vec<Flower>) -> Vec<Data> {
-    let mut input_data = Vec::with_capacity(input.len());
-    for i in 0..input.len() {
-        input_data.push(structs::Data::from_flower(input[i]));
-    }
-    input_data
-}
+
 
 // TODO: DOK
 fn parse_data<T>(datafile: &str) -> Result<Vec<T>, io::Error>
